Refactored for Replace Conditional with Polymorphism & Pull-up variable/method
