+++ src/main/java/org/isf/serviceprinting/manager/PrintManager.java
@@ -65,7 +65,7 @@ public class PrintManager {
 	
 	public PrintManager() {}
 	
-	public void print(String filename, List<?> toPrint, int action) throws OHServiceException {
+	public void print(String filename, List<?> toPrint, Action action) throws OHServiceException {
 		
 		Map<String, Object> parameters = new HashMap<>();
 		Hospital hospital = hospitalManager.getHospital();
@@ -84,31 +84,7 @@ public class PrintManager {
 				.loadObject(jasperFile);
 				JasperPrint jasperPrint = JasperFillManager.fillReport(
 						jasperReport, parameters, dataSource);
-				switch (action) {
-				case 0:
-					if (GeneralData.INTERNALVIEWER) {
-						JasperViewer.viewReport(jasperPrint,false, new Locale(GeneralData.LANGUAGE));
-					} else {
-						String pdfFile = "rpt_base/PDF/" + filename + ".pdf";
-						JasperExportManager.exportReportToPdfFile(jasperPrint, pdfFile);
-						try {
-							Runtime rt = Runtime.getRuntime();
-							rt.exec(GeneralData.VIEWER +" "+ pdfFile);
-						} catch(Exception exception) {
-							LOGGER.error(exception.getMessage(), exception);
-						}
-					}
-					break;
-				case 1:
-					JasperExportManager.exportReportToPdfFile(jasperPrint,"rpt_base/PDF/"+
-							JOptionPane.showInputDialog(null,MessageBundle.getMessage("angal.serviceprinting.selectapathforthepdffile.msg"), filename)
-							+".pdf");
-					break;
-				case 2:JasperPrintManager.printReport(jasperPrint, true);
-					break;
-				default:JOptionPane.showMessageDialog(null,MessageBundle.getMessage("angal.serviceprinting.selectacorrectaction.msg"));
-					break;
-				}
+				action.Print(jasperPrint, filename);
 			} else {
 				JOptionPane.showMessageDialog(null,MessageBundle.getMessage("angal.serviceprinting.notavalidfile.msg"));
 			}
